// iot-service/iot-angular/src/sdk/core/models/node-profile-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NodeProfileResponseDto {
  code: string;
  createdAt: string;
  description?: string;
  enabled: boolean;
  idNodeModel: string;
  idNodeProfile: string;
  idProject?: string;
  mappingJson: {
};
  name: string;
  nodeCount?: number;
  nodeModel?: {
};
  parserType: string;
  project?: {
};
  transformScript?: string;
  updatedAt: string;
}